algo_server/pyproject.toml
[tool.poetry]
name = "algo-server"
version = "0.1.0"
description = "AlgoCoach Server"
authors = ["Emma Chen <“echen053@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.0"
starlette = "^0.35.1"
openai = "^1.10.0"
langchain = "^0.1.4"
langchain-openai = "^0.0.5"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
